requirements.txt
fastapi
uvicorn
ollama
pyht
httpx
pymongo
python-dotenv
openexr
Pillow
numpy